requirements.txt
openai
openai-whisper
gTTS
pydub
python-dotenv
streamlit
pyaudio